requirements.txt
streamlit>=1.33.0
crewai>=0.30.0
python-dotenv>=1.0.1
PyPDF2>=3.0.0
matplotlib>=3.8.0